showing performance time as float with two decimals and generating less random count object to allow to generare about 100 objects when is called with -c 100

+++ src/DTOPerformanceComparison/DTOs/OwnDTO/ResultDTO.php
@@ -7,7 +7,7 @@ namespace DTOPerformanceComparison\DTOs\OwnDTO;
 class ResultDTO
 {
     public function __construct(
-        public readonly int $timeSpent,
+        public readonly float $timeSpent,
         public readonly int $memoryUsed,
     ) {
     }
